update paths to common.py (renamed from: data.py)
