+++ frontend/pages/forecasts.py
@@ -11,7 +11,7 @@ from components.layout import Row, Column, Container
 from components.lists import ListSimple
 from components.table import SimpleTable
 from components.controls import SubmitButton
-from pages.data.data import (
+from pages.data.common import (
     username,
     year_month_dict_list,
 )
