+++ frontend/pages/rates.py
@@ -12,7 +12,7 @@ from components.layout import Row, Column, Container
 from components.lists import ListSimple
 from components.table import SimpleTable
 from components.controls import SubmitButton
-from pages.data.data import (
+from pages.data.common import (
     months_start,
     username,
 )
